requirements.txt
fastapi
uvicorn
openai
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
PyPDF2
python-dotenv
aiofiles
